Add headers and modularised communications
- xbee_driver is functions relating to talking to the xbee
- uart_interface sets up the uart peripheral and provides write functions and interrupts
- renamed functions from communications to more conventional name format (camel case and module prefix on all function names)
- renamed timer0 to timer_interface
