+++ RobotNoASF/Functions/manual_mode.h
@@ -30,7 +30,7 @@
 #define CCW 0xD3
 
 //////////////[Functions]///////////////////////////////////////////////////////////////////////////
-void manualControl(struct message_info tmessage, struct Position *robotPosition);
+void manualControl(struct MessageInfo tmessage, struct Position *robotPosition);
 
 
 #endif /* MANUAL_MODE_H_ */
