+++ RobotNoASF/robot_setup.h
@@ -71,9 +71,10 @@ struct Position
 #include "sam.h"
 #include "Interfaces/pio_interface.h"
 #include "Interfaces/imu_interface.h"
-#include "Interfaces/timer0.h"
+#include "Interfaces/timer_interface.h"
 #include "Interfaces/external_interrupt.h"
-#include "Interfaces/communication.h"
+#include "Interfaces/uart_interface.h"
+#include "Interfaces/xbee_driver.h"
 #include "Interfaces/adc_interface.h"
 #include "Interfaces/opt_interface.h"
 #include "Interfaces/motor_driver.h"
