+++ RobotNoASF/swarm_robot_V2.cproj
@@ -246,10 +246,10 @@
     <Compile Include="Interfaces\adc_interface.h">
       <SubType>compile</SubType>
     </Compile>
-    <Compile Include="Interfaces\communication.c">
+    <Compile Include="Interfaces\xbee_driver.c">
       <SubType>compile</SubType>
     </Compile>
-    <Compile Include="Interfaces\communication.h">
+    <Compile Include="Interfaces\xbee_driver.h">
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="Interfaces\external_interrupt.c">
@@ -309,16 +309,22 @@
     <Compile Include="Interfaces\spi.h">
       <SubType>compile</SubType>
     </Compile>
-    <Compile Include="Interfaces\timer0.c">
+    <Compile Include="Interfaces\timer_interface.c">
       <SubType>compile</SubType>
     </Compile>
-    <Compile Include="Interfaces\timer0.h">
+    <Compile Include="Interfaces\timer_interface.h">
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="Interfaces\twimux_interface.c">
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="Interfaces\twimux_interface.h">
+      <SubType>compile</SubType>
+    </Compile>
+    <Compile Include="Interfaces\uart_interface.c">
+      <SubType>compile</SubType>
+    </Compile>
+    <Compile Include="Interfaces\uart_interface.h">
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="robot_setup.c">
